chore: adjust some fix based on review
